feat(simulator): split simulator code
