+++ src/components/Simulator/init/dom.js
@@ -161,134 +161,4 @@ export default function initDom(interpreter, globalObject) {
     "createTextNode",
     interpreter.createNativeFunction(wrapper),
   );
-
-  // Define 'alert()' function.
-  var wrapper = function alert(text) {
-    return window.alert(arguments.length ? text : "");
-  };
-  interpreter.setProperty(
-    globalObject,
-    "alert",
-    interpreter.createNativeFunction(wrapper),
-  );
-
-  // defined console.log
-  var wrapper = function log(text) {
-    console.log(text);
-  };
-  interpreter.setProperty(
-    globalObject,
-    "log",
-    interpreter.createNativeFunction(wrapper),
-  );
-
-  // ------------------------------------------------------------
-  // TIME
-  // ------------------------------------------------------------
-
-  // Define 'delay(number)' function.
-  var wrapper = function (ms, next) {
-    window.setTimeout(function () {
-      next();
-    }, ms);
-  };
-  interpreter.setProperty(
-    globalObject,
-    "delay",
-    interpreter.createAsyncFunction(wrapper),
-  );
-
-  // ------------------------------------------------------------
-  // OLED DISPLAY
-  // ------------------------------------------------------------
-
-  // Define 'drawText' function.
-  var wrapper = function (text, startX, startY, size, color) {
-    const SCALE = 2;
-    const baseTextSize = 6;
-    const textSize = baseTextSize * SCALE * size;
-    const canvas = document.getElementById("oled-display");
-    const ctx = canvas.getContext("2d");
-
-    const x = startX * SCALE;
-    const y = startY * SCALE + textSize;
-
-    if (color === "BLACK") {
-      // Measure text dimensions
-      const textMetrics = ctx.measureText(text);
-      const textWidth = textMetrics.width;
-      const textHeight = textSize; // Approximation based on font size
-
-      // Draw black box behind the text
-      ctx.fillStyle = "white";
-      ctx.fillRect(x, y - textHeight, textWidth, textHeight + 4);
-
-      // Draw the text
-      ctx.fillStyle = "black";
-      ctx.font = `${textSize}px monospace`;
-      ctx.fillText(text, x, y);
-    } else {
-      // Draw the text
-      ctx.fillStyle = "white";
-      ctx.font = `${textSize}px monospace`;
-      ctx.fillText(text, x, y);
-    }
-  };
-
-  interpreter.setProperty(
-    globalObject,
-    "drawText",
-    interpreter.createNativeFunction(wrapper),
-  );
-
-  // Define 'display.clearDisplay' function.
-  var wrapper = function clearDisplay() {
-    const canvas = document.getElementById("oled-display");
-    const ctx = canvas.getContext("2d");
-    ctx.fillStyle = "black";
-    ctx.fillRect(0, 0, canvas.width, canvas.height);
-  };
-  interpreter.setProperty(
-    globalObject,
-    "clearDisplay",
-    interpreter.createNativeFunction(wrapper),
-  );
-
-  // ------------------------------------------------------------
-  // TEMPERATURE SENSOR
-  // ------------------------------------------------------------
-
-  // Define getTemperature function
-  var wrapper = function readTemperature() {
-    return document.getElementById("temperature-slider").value;
-  };
-  interpreter.setProperty(
-    globalObject,
-    "readTemperature",
-    interpreter.createNativeFunction(wrapper),
-  );
-
-  // Define getTemperature function
-  var wrapper = function readHumidity() {
-    return document.getElementById("humidity-slider").value;
-  };
-  interpreter.setProperty(
-    globalObject,
-    "readHumidity",
-    interpreter.createNativeFunction(wrapper),
-  );
-
-  // ------------------------------------------------------------
-  // NEOPIXEL
-  // ------------------------------------------------------------
-  // Define neopixel function
-  var wrapper = function neopixel(r, g, b) {
-    document.getElementById("board-complex_svg__circle270").style.fill =
-      `rgb(${r}, ${g}, ${b})`;
-  };
-  interpreter.setProperty(
-    globalObject,
-    "neopixel",
-    interpreter.createNativeFunction(wrapper),
-  );
 }
