+++ src/components/Simulator/init/index.js
@@ -1,0 +1,15 @@
+import initDisplay from "./display";
+import initDom from "./dom";
+import initLogAndAlert from "./log";
+import initNeopixel from "./neopixel";
+import initHDC1080 from "./sensors/hdc1080";
+import initTime from "./time";
+
+export default function initSimulator(interpreter, globalObject) {
+  initDom(interpreter, globalObject);
+  initLogAndAlert(interpreter, globalObject);
+  initTime(interpreter, globalObject);
+  initDisplay(interpreter, globalObject);
+  initHDC1080(interpreter, globalObject);
+  initNeopixel(interpreter, globalObject);
+}
